Link the cabal file inside the cabal2nix build to resolve #5
cabal2nix needs to be given a directory containing a cabal
file (giving it a file instead of a directory changes what it does).

This change links the cabal file into the $out directory so cabal2nix
will only see the contents of the $out directory as opposed to looking
at the entire Nix store.