Update nixpkgs, remove static binaries and overrides (#21)
With a fresh nixpkgs release ahead of us (21.05) we automatically get
the latest versions of all interactive tools.  That means we can clear
out the existing overrides.

This commit also removes the ability to build static binaries as
advertised in the last release.

The tests are failing for GHC 9.0.1.  I need to clean up the code a
bit before I address that.